% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countES.R
\name{countES}
\alias{countES}
\title{Calculates effect size measures for count regression models}
\usage{
countES(
  int,
  int_se,
  slope,
  slope_se,
  disp,
  mtype,
  reps,
  CI_level,
  randseed,
  eff_plot,
  dist_plot
)
}
\arguments{
\item{int}{Intercept value (numeric)}

\item{int_se}{Standard error of the intercept (numeric)}

\item{slope}{Slope value (numeric)}

\item{slope_se}{Standard error of the slope (numeric)}

\item{disp}{Dispersion: square root of phi for overdispersed, alpha for negative binomial (numeric)}

\item{mtype}{Model type ("poisson" "overdispersed" "negative binomial")}

\item{reps}{Number of Monte Carlo replications (integer)}

\item{CI_level}{Confidence interval level (e.g., 95 for 95\% confidence interval)}

\item{randseed}{Random seed to replicate results (integer)}

\item{eff_plot}{Plot of exponential effect (TRUE FALSE)}

\item{dist_plot}{Histogram of Monte Carlo replications (TRUE FALSE)}
}
\description{
This function calculates the standardized mean difference (SMD, i.e., Cohen's d)
and the multiplicative effect size for a count regression model
}
\examples{

# Poisson regression
countES(int = 1, int_se = 0.5, slope = 0.2, slope_se = 0.1,
disp = 0, mtype = "poisson",
reps = 10000, CI_level = 95, randseed = 12345,
eff_plot = TRUE, dist_plot = TRUE)

# Overdispersed Poisson regression
countES(int = 1, int_se = 0.5, slope = 0.2, slope_se = 0.1,
disp = 1.5, mtype = "overdispersed",
reps = 10000, CI_level = 95, randseed = 12345,
eff_plot = TRUE, dist_plot = TRUE)

countES(int = 1, int_se = 0.5, slope = 0.2, slope_se = 0.1,
disp = 1.5, mtype = "negative binomial",
reps = 10000, CI_level = 95, randseed = 12345,
eff_plot = TRUE, dist_plot = TRUE)

countES()
}
\keyword{count}
\keyword{effectsize}
\keyword{montecarlo}
\keyword{negativebinomial}
\keyword{poisson}
